+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/ajax/dto/ajax/AjaxErrorResponse.java
@@ -1,9 +1,9 @@
 package ca.corefacility.bioinformatics.irida.ria.web.ajax.dto.ajax;
 
 /**
-￼ * Response returned if their is an error creating a new item in IRIDA.
-￼ * Will return the internationalized explanation for the error.
-￼ */
+ * Response returned if there is an error creating a new item in IRIDA.
+ * Will return the internationalized explanation for the error.
+ */
 public class AjaxErrorResponse extends AjaxResponse {
 	private final String error;